backend/requirements.txt
google-generativeai
feedparser
beautifulsoup4
requests
python-dotenv
html2text
